Encapsulation services + Correction deletes + VM

+++ Digiway/DigiwayWebapi/Controllers/ValuesController.cs
@@ -18,7 +18,7 @@ namespace DigiwayWebapi.Controllers
 
         // GET api/values/5
         [HttpGet("id/{id}")]
-        public string Get(int id)
+        public string Get(long id)
         {
             return "value";
         }
@@ -31,13 +31,13 @@ namespace DigiwayWebapi.Controllers
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody]string value)
+        public void Put(long id, [FromBody]string value)
         {
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public void Delete(long id)
         {
         }
     }
